Fix and improve US 7-Eleven spider
The most recent change to this spider switched from hitting an API endpoint to scraping data from the HTML directly. Enough things had broken in the HTML structure, that I switched it back to hitting the API.

The API endpoint seems to be protected in two ways:

1. It requires an OAuth Bearer token to access
For anonymous sessions, this token is fetched using a static set of client credentials before first use. It's possible that this client ID / secret is rotated regularly with website deploys, but I haven't seen any evidence of that in a few days of testing.

2. It requires a set of X-SEI-* headers on every request.
These headers seem to be used in tandem to validate the request. The X-SEI-TRIP-ID appears to be a base64-encoded hash of info from the user agent, time zone, platform, and version. From skimming the minified JS code that generates it, there doesn't appear to be any obvious time-based component that would case it to expire, so the static credentials here are likely good unless the algorithm changes or old versions get locked out.