fix londis/premier spiders with respect to websites changing, fix lint errors
so I had been running my code against cached copied of the londis and premier sites for the last few weeks.
the websites had in the meantime changed HTML on me. just goes to show that tests are needed on historic
output dumps, is this weeks run very diffeent from last weeks?