Add spider for Western Union
This one is slow to start as it starts up and immediately enqueues half
a million scrapy Request objects in memory. Supplying
`scrapy crawl western_union -s JOBDIR=western_union-job/`
allows it to flush the Request objects to disk until they are needed and
reduces the amount of heap it needs. Might not be an issue on a machine
with sufficient RAM.

The site is happy to serve 800 pages/min, completes in under 8 hours,
produces 393778 items.