Extend the dictionary of dict parser
The dict parser will need to have its dictionary extended as we move forward.
Based on the experience of using a similar function in my own project I have
contributed what I consider a reasonable dictionary. It is based on the
experience of hundreds of spiders and shoudl save a bit of work going forward.

The separation of the data diciotnary from the logic I also think makes
life clearer.