Fix Ashley HomeStore (sic) spider
Decided to "fix" as it was a large file. Uncovered a numbert of problems:

Ashley HomeStore is in Canada.
Ashley Furniture in US ... and in fact beyond.

NSI does not support Ashley Furniture outside US ... oopsy?
Hello better tooling support for synergies between ALP / NSI / wikidata / OSM?

Anyway two spiders as one is sitemap and the other cannot be.

End result of "clean" is we go from 738 (US only!!) stores to
approx 1K global.