some doc about spider naming convention
The top level README will get too unwieldy for some of the things we
should be saying (we have quite a few things to say). I think we should
break things down into little vignettes that we can tie together in the
main README or wherever.

The files will serve as reference but also single links that can
be pasted into a PR to help the reviewer get a point across in
a particular area. Who likes typing? :-)