add a docs page about wikidata/nsi/atp
Cleaned up a little bit my previous SPIDER_NAMING commit. In particular made
the linked spiders as best as they could be given that people would be looking
at them!

The new document page is a first stab at the subject and has a fuller description
of the "scrapy nsi" command than is in the top level README.