Emit properties in order.
We were treating pending as a stack; use it as a queue. The algorithm
says to sort results in tree order, which I'm still ignoring, but this
is more likely to produce the properties in the same order they came in.

(Ref: https://html.spec.whatwg.org/multipage/microdata.html#associating-names-with-items)