Change format of weekly dump files to `*.zip` instead of `*.tar.gz`
The size of the compressed dump is about the same: 146M vs. 147M with
the current dump, which isn’t surprising given that both formats use
the same compression algorithm. But other than with tar, the files
in a ZIP archive can be accessed separately without having to read
the entire stream. This property of ZIP allows processing the weekly
dump files in parallel with multiple threads/processes/machines.

Closes https://github.com/alltheplaces/alltheplaces/issues/4227.