Sort same-day opening hours
There had already been a unit test for sorting multiple opening hours
on the same day, but the implementation had not actually sorted the hours.
Before this change, the results thus got formatted in pseudo-random order
depending on the value of Python’s built-in hash function. Accidentally,
the existing test case happened to give the expected sorting order.