Add Prada (Global) (373 stores)
Note that this spider currently extracts 373 stores, but the store
finder on the website indicates only 368 stores exist. There could be a
few recently closed stores that are being extracted from the sitemap? It
isn't clear yet why the difference exists and if it were due to recently
closed stores, how these stores could be enumerated and removed from the
sites extracted.

The street_address field extracted from linked data is very messy.
Sometimes it really is street_address, sometimes it is full_addr, but
most of the time it is neither and is a street_address and state
combined in a myriad of creative ways. Many street_address fields are
completely in other languages such as Simplified Chinese or Korean.

Closes issue #4721