bunnings: fix spider and include New Zealand stores
The Bunnings API used for locating stores requires a guest/anonymous
OAUTH login process to be completed, and an Authorization Bearer header
to be sent when calling the API.

The spider has largely been rewritten using DictParser, fixing numerous
incorrect field values in the process.

The rewritten spider also extracts New Zealand Bunnings stores.

Currently 351 stores are returned by the API, with 41 being in New
Zealand and the remainder in Australia.