OpeningHours: introduce add_ranges_from_string (#4893)
* OpeningHours: introduce add_ranges_from_string

add_ranges_from_string is added as a parser function designed to help
translate from opening hours provided as human readable text in a fuzzy
format.

This function is intended to add opening hours ranges based on the
following type of ranges that may be found in source data:

1. Monday-Wednesday: 5pm - 7pm
2. Monday-Wednesday 08:00-14:00
3. Monday to Tuesday: 15:00:01 to 16:35
4. Thurs 2PM-6:30PM
5.  Fri    9am  -  11am
6. Weekends: 8:00 AM to 6:00 PM
7. Monday to Thursday 7am to 7pm, Friday 7am to 10pm, Weekends CLOSED

This function works with a source that may mix and match from the type
of examples listed above. For example, if a source uses both 12h and 24h
times to describe opening hours.

This function does not currently support extraction of multiple opening
hours on the same day of a week. Only the first opening time
of each day is extracted. This function should therefore not be used if
the source data has the potential to list multiple opening hours for the
same day.

This function does not currently support named times of each day such as
'Midday' and "Midnight". It is possible to replace these prior to the
input string being provided to this function.

This function allows specification of days and day ranges in langauges
other than the default English.

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* OpeningHours: improve add_ranges_from_string

Add a named day for "Daily" for all days of the week.

Capture date ranges that commence on a Sunday e.g. Sunday-Thursday as
some sources have their hours ranges commence on Sunday instead of the
typical start day of Monday.

Allow times to have no delimiter between HH and MM e.g. 2000 instead of
20:00.

Improve documentation of the add_ranges_from_string function.

* OpeningHours add_ranges_from_string: improve and add ES tests

Improve add_ranges_from_string with an ability to customise
delimiters on a per-language basis.

Add Spanish tets for use of add_ranges_from_string.

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* OpeningHours add_ranges_from_string: add more EN delimiters and ranges

* OpeningHours add_ranges_from_string: improve extraction

* Improve 12h time extraction so that if AM/PM is missing, make an
  assumption about what may have existed.

* Improve day/named day range extraction so that it is fully case
  insensitive.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>