Add SweetIQ store locator
This store locator is similar to the UberAll store locator in that
SweetIQ acquired UberAll. Store locator IDs for the SweetIQ spider can
often be used as the API key for a UberAll spider. However, the SweetIQ
store locator API has more fields of information, and would appear to be
the future/replacement of the UberAll store locator.