Add SweetIQ store locator and sample brands which use this store locator (#5058)
* Add SweetIQ store locator

This store locator is similar to the UberAll store locator in that
SweetIQ acquired UberAll. Store locator IDs for the SweetIQ spider can
often be used as the API key for a UberAll spider. However, the SweetIQ
store locator API has more fields of information, and would appear to be
the future/replacement of the UberAll store locator.

* Add Cashco Financial (CA) (57 stores)

* Add First Hawaiian Bank (US/GU) (51 branches and 180 ATMs)

* Add Golf Town (CA) (48 stores)

* Add Quesada (CA) (188 stores)

* Add Sizzler (US/PR) (84 stores)

* Add Sporting Life (CA) (14 stores)

* Add St. Louis Bar & Grill (CA) (74 stores)

* Add The Paper Store (US) (94 stores)

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* storefinders/sweetiq: address review comments to improve spider

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>