fantastic_furniture_au: fix spider
A public holiday this week in Australia has demonstrated that opening
hours for Fantastic Furniture are sometimes changed to align with public
holidays due in the next week. In this case, Friday is "CLOSED", so the
spider needs to be able to accomodate these temporary changes to hours.
Unfortunately there is no information available to determine whether
hours are normal or holiday hours, so the opening hours for this spider
will seemingly fluctuate regularly.