jax_tyres_and_auto_au_fix: fix spider
With upcoming public holidays in Australia this week, this spider is
broken due to unhandled day names such as "Good Friday". Unfortunately
for such weeks with a public holiday, there is no place on the website
showing what the usual hours would have been. Thus the hours for this
spider will constantly fluctuate depending on whether the given week has
a public holiday.