storefinders/where2getit: rewrite to use getlist API call
The Where2GetIt store finder was using an inefficient geolocation search
API call that can be significantly improved and simplified by use of a
different getlist API call that returns all locations, not just those in
a geographic radius of a point.

For particularly large brands, this rewritten store finder allows
locations to be returned on a country-by-country basis to avoid the
server timing out in its response to the getlist API call.

Fixes #5119