freddys: delete as broken and behind CloudFlare bot protection
This spider is broken and can't be fixed (at least as present because
ATP doesn't currently try to use external tools to pretend to be a full
fledged web browser). In the future perhaps tools such as
https://github.com/FlareSolverr/FlareSolverr would be integrated with
ATP to allow this spider to be added again.

Reference: https://github.com/alltheplaces/alltheplaces/pull/5184