western_union: fix spider to use correct geographic search
The Western Union GraphQL endpoint does use a 350km radius from each
supplied search coordinate to find locations to return (over multiple
pages of results). The previous attempt to fix this spider incorrectly
assumed a 'country' parameter to the GraphQL search function was used
instead of a radius being applied.

To achieve this search across the entire surface of the Earth, a new
searchable points file is provided, along with documentation for how
this searchable points file was created, and how similar could be
created in the future for different search radiuses.