action: fix spider
Whilst a GraphQL API still exists, it has changed and no longer provides
in the list of all locations the required IETF language tag (e.g. fr-FR)
for each location. Some language variants of the Action website use a
much simpler API that returns the same information previously captured,
just without the fuss of having to reverse geocode coordinates and map
countries to IETF language tags. Language variants of the Action website
which use a GraphQL based store locator are broken because whilst they
show a marker for a location in every country, details of locations in
countries not using the same IETF language tag as the store locator
language variant will not load.