ci/run_all_spiders.sh: save end time of each build to history.json
Currently, the start time of a build is saved in history.json. Also
saving the end time of a build allows for statistics to be more easily
captured for how long it is taking to complete each complete build. With
this data, it may then be possible to see the effect of efforts to
optimise spider crawls to reduce overall build times.