Fix FedEx. (#5616)
* Fix FedEx.

Test me with `scrapy sd https://local.fedex.com/pl-pl/zyrardow/wawek`.

The microdata contains nested itemprop=name, which gets translated as an
array of names. Teach LinkedDataParser to use the first name when a list
is found.

When an item contains an array of names, an exception is thrown. Teach
ClosePipeline to guard against this so that it does not cause items to
be lost.

The `scrapy sd` command prints out the found item, then a message saying
"failed to decode structured data", which does not explain anything.
Change its log level from disabled to warning so that exceptions are not
swallowed but it otherwise remains terse.

Sort the stats output from `scrapy sd`.

Fix isinstance() call in CheckItemProperties.

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Update locations/commands/sd.py

Co-authored-by: Cj Malone <Cj-Malone@users.noreply.github.com>

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Cj Malone <Cj-Malone@users.noreply.github.com>