Consolidate Mosaic Brands spiders and fix
A number of brands owned by Mosaic Brands have changed sitemaps,
requiring a different approach to crawling locations.

This is an opportunity to consolidate the various Mosaic Brands spiders
together with a common approach to crawling locations from a store
directory page on each brand website.

W Lane (w_lane_au) does not have a store directory page nor a suitable
sitemap, and the only alternative appears to be:
https://www.wlane.com.au/on/demandware.store/Sites-mosaic-au-Site/en_AU/Stores-FindStoresAvailablity?postalCode=2000&radius=10000&pids=SKU0

This API call returns 886 locations across all of Mosaic brands,
including numerous locations with "CLOSED" in the name and URL. Some
closed locations such as "Crossroads" brand locations are not marked
"Closed" in the response to this API call, and indicates this API call
is returning both open and closed locations (not marked as such) and
potentially stale data. The API call is not available on other brand
websites, and thus it appears this could be an old or soon to be
replaced API call.

For now, W Lane's old (broken) spider is removed without a replacement.
If the W Lane website is updated with a store directory like the other
brands of Mosaic Brands, W Lane can be added to this Mosaic Brands
spider again.