decathlon: fix spider
Decathlon implements the Woosmap store finder for their larger
(generally European) websites. There is a separate API key per country
that Decathlon operates within. This change removes the old Decathlon
spider and replaces it with country-specific WoosmapSpider spiders. No
locations that would have been captured by the old "decathlon" spider
would be missed with the new country-specific WoosmapSpider spiders.

A new Python dependency "unidecode" is also introduced as Decathlon's
website URLs are an ASCII-equivalent transliteration of the store name,
and "unidecode" provides the ability to convert European language
diacritics to the closest ASCII-equivalent character.

There are other countries within which Decathlon operates however they
appear to have only a few stores each, and the websites do not use a
country-specific Woosmap store finder API key. For most of these other
countries, it is possible that each store page needs to be parsed
individually to extract store location information. This is separate
work unrelated to the replacement of the old "decathlon" spider with
country-specific WoosmapSpider spiders.