storefinders/woosmap: add origin as a required class attribute (#5673)
* storefinders/woosmap: add origin as a required class attribute

Woosmap always requires an "Origin" HTTP header to be provided. Update
the store finder to require an "origin" attribute be supplied, that is
then sent to the Woosmap API along with the API key for the brand.

Woosmap additionally appears to allow opening hours to be specified as
"default" e.g. hours applying across all days of the week. Allow this
"default" hours specification method to be used.

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>