sainsburys: fix spider (#5678)
At least one location did not have the "other_name" field populated,
meaning that the name of the store was being extracted as a None value,
causing a fatal error due to invalid handling of the None value. The
number of stores returned by this spider has therefore been incorrect,
as results have been truncated.

This change will only use the "other_name" field if it contains a value,
otherwise, the abbreviated "name" field will be used.