costacoffee_gb: fix and rename to costa_coffee_gg_gb_im_je (#5717)
* costacoffee_gb: fix and rename to costa_coffee_gg_gb_im_je

The store locator has been switched to GraphQL.

Note that the GraphQL query used is a "closest N locations to
coordinates X,Y" query where N can be any large number. If N is too
large, the GraphQL endpoint times out.

We thus need to split up the GraphQL query search into smaller search
areas. A density of 2500 locations per 100km radius seems to return all
locations whilst avoiding some duplication. This is not a perfect way to
search because the density in Northern Scotland is quite low, and the
nearest 2500 locations are still returned. For London with a very high
density, a density of 2500 locations per 100km radius seems to be
necessary. Due to the inability for this spider to currently control for
density of search areas, there is a high number of duplicate item
warnings.

* costa_coffee_gg_gb_im_je: add DGGRID metafile

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* costa_coffee_gg_gb_im_je: fix DGGRID metafile documentation

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Ian Dees <ian.dees@gmail.com>