storefinders/store_locator_plus: split into cloud vs. self-hosted
StoreLocatorPlus is both a software-as-a-service store finder, as well
as an older self-hosted WordPress plugin. This change introduces a new
StoreLocatorPlusSelfSpider for self-hosted, and renames the existing
StoreLocatorPlusSpider to StoreLocatorPlusCloudSpider.

hemmakvall_se uses the new StoreLocatorPlusSelfSpider.

ruler_foods_us is a new spider that uses StoreLocatorPlusSelfSpider.

national_seating_mobility uses StoreLocatorPlusCloudSpider.

Fixes #5844