Add Specsavers (GB/CA/AU) (2611 locations)
 'atp/category/healthcare/audiologist': 1208,
 'atp/category/healthcare/optometrist': 1403,

Removes specsavers_au_nz spider which was broken. New Zealand and
European locations for Specsavers use a different store locator (not
based on Sparql queries).

Partially reverses #5319 now that the store locator for some countries
has changed to Sparql, and the Cloudflare blocking of the past is no
longer in place.