Enable AutomaticSpiderGenerator for MetaLocatorSpider
MetaLocatorSpider detection requires the use of Playwright to wait for
Magento to dynamically create an iframe on the store locator page, where
the iframe src attribute is to code.metalocator.com and contains the
brand identifier needed to generate a new spider based on
MetaLocatorSpider automatically.

It appears that specifying a country list for MetaLocatorSpider is not
required, so this has been removed.