Add Marathon Petroleum (US) to replace marathon and arco
Marathon Petroleum owns the Marathon and Arco brands, and the same store
finder is used for both brands. Currently the `marathon` and `arco`
spiders are broken. These old spiders are now replaced by the single
Marathon Petroleum spider this commit introduces.

Fuel types are very confusing and there are 100's of different codes
applying to different types of fuels. The matching between codes and OSM
fuel tag values is best-effort.