Add Vertiche (MX) (246 stores) (#7323)
* Add Vertiche (MX) (246 stores)

This is a complicated variant of WPStoreLocatorSpider that requires
offline lookup of city names to coordinates. It seems to have most, if
not all, locations of Vertiche being returned. There is a chance,
particularly over time, that this spider needs updating further to
manually map more city names on the store locator page to geonamescache
city names, or perhaps even to coordinates that Google Maps APIs would
have returned.

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>