Add Anti-Bot Detection middleware
This new middleware is in two parts:

1. Middleware to detect use of anti-bot methods such as Cloudflare
   Captchas, and various other Web Application Firewalls (WAF)
   implementing the likes of TLS fingerprinting, browser JS challenges,
   etc.

2. Middleware to stop a spider from generating more requests, leading to
   quick (but still deferred) stopping of a crawl once anti-bot methods
   have been detected.

Spiders can implement a new attribute "anti_bot_methods" which is a list
of AntiBotMethod as specified in
locations/middlewares/anti_bot_detection.py

AntiBotDetectionMiddleware will automatically set the attribute if
anti-bot method(s) are discovered.

If a particular anti-bot method is unexpected or if Zyte API is not
available to use for bypassing the anti-bot method, AntiBotStopCrawl
will automatically shut down the crawl in a graceful way, generally
after all requests already sent have been received and parsed.