DictParser: fix reference keys
Commit e3ecda5778fdff7ba8261ca4fd3a2cc678abfde7 tried to standardise
case of reference keys. The idea was that the variation function would
try all possibilities of case for each key (camel case, title case,
uppercase, lowercase, etc).

Unfortunately it was forgotten that the "ID" acronym in reference keys
cannot be handled well by the variations function. It is common for a
reference key to always capitalise acronyms, particularly "ID", whilst
using some other case for everything else.

This change should fix a number of spiders that were returning 0 or 1
result due to lack of ability to set unique references for each location
scraped.