Add City of Darwin CCTV (106 cameras)
This is all fixed (non PTZ) cameras. The coordinates correspond to
poles where PTZ cameras also appear to be mounted, even though a feature
is not able to be extracted for each PTZ camera.

Compare with `CameraCount` field of
https://www.arcgis.com/apps/mapviewer/index.html?url=https://services6.arcgis.com/tVfesLETUHNU9Vna/ArcGIS/rest/services/CCTV_Cameras/FeatureServer&source=sd
to see where PTZ cameras are likely to exist.

One *could* extract PTZ camera locations but it would require somewhat
complex queries of the `CameraCount` field in the `CCTV_Cameras` layer
versus counting camera points from the `CCTV_Points` layer nearby to
each point in the `CCTV_Cameras` layer, showing differences generally of
1 PTZ camera here and there.