Add WinMart and WinMart+ (VN) (2422 stores)
Source data is a bit patchy--for example, some locations do not have
coordinates provided.

There are other alternative APIs available, all requiring different
forms of search and with different fields of data returned. The selected
API of this spider is the only one found to return all stores in a
single query, and seems to contain most data fields that are of
interest.

Fixes #9103