Add Pedders Suspension & Brakes (shop/car_repair) (AU) (80 stores) (#11778)
* Add Pedders Suspension & Brakes (shop/car_repair) (AU) (80 stores)

Note: it is not yet known how the website determines whether a location
is an official store or authorised dealer. A less-than-perfect method is
currently used to estimate whether a location is an official store.
Thus 80 stores are returned (of 121 stores total) but some of these 80
stores are authorised dealers. Ideally these authorised dealers would be
ignored.

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* pedders_suspension_and_brakes_au: fix regex escaping

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>