Add Rosetta APR "storefinder" for Australian electricity features
Rosetta APR is used by most Australian energy grid operators for
displaying maps of electricity grid assets such as substations and
lines. Rosetta APR and this type of data is generally made publicly
available.

Depending on the energy grid operator, different types of files may be
served by the Rosetta APR platform such as geojson files (for features
with coordinates) and csv or zipped csv spreadsheet files for detailed
data on features such as capacities of substations or transformers. This
storefinder provides a means to make second and subsequent requests for
additional data files to allow extracted features to be supplemented
with more detailed data such as capacity ratings of transformers.