Add Powercor (power/(substation|transformer)) (AU) (93779 features)
Powercor and Citipower are separate electricity grid operators but have
common ownership and common ICT systems. As a result, this spider
combines assets such as substations and transformers under the Powercor
brand only, as there is no method in the source data for determining
whether an asset is a Powercor asset or Citipower asset.

Feature types extracted:
 'atp/category/power/substation': 97,
 'atp/category/power/transformer': 93682,