chris_willemse_cycles_za: fix spider
Only one location left and no valid coordinates are available. I was
going to delete this chain but it is part of a broader group and
therefore even a single location may be of interest. I started working
on it before realising the coordinates are invalid. Oh well. Here is one
single shop with a seemingly valid address and phone number.