settings: refactor and remove circular dependency with geojson exporter
Most of the Scrapy settings which accept class references as strings
also accept the class type itself. The difference is lazy loading of
extensions, pipelines and exporters (and similar) with class references
as strings, versus eager loading of these classes when Scrapy is
started. Almost everything of substance that settings.py was going to
lazy load is eventually loaded in every ATP crawl anyway, so lazy
loading provides little to no benefit. _IF_ there was an expensive
pipeline in the future which took considerable time to load then it
would make better sense to selectively load it using a spider's
overridden SETTINGS attribute, rather than apply that pipeline to every
single item from every spider.

For the few classes settings.py called out to (such as exporters) which
would typically not always be loaded, these are extremely simple in
nature and the performance benefit of lazy loading these simple classes
will be negligable.

The benefit of eager loading from settings.py is the ability to more
readily detect syntax errors in extensions, pipelines and exporters
(and similar) without having to wait for a complete crawl to complete,
and without having to attempt a crawl of each combination of extensions,
pipelines and exporters (and similar) to ensure all are free of syntax
errors.

To enact this change, extensions, pipelines and exporters (and similar)
can no longer import from settings.py as this would create a circular
redundancy. The GeoJSON exporter was unnecessarily importing from
settings.py and this has had to be removed in favour of an alternative
method that doesn't create a circular dependency.