Update function documentation for handling ambiguous data
Per discussion in PR #12568.

If source data has two features:

{
  "feature_id": "1",
  "services": ["Showers", "Toilets"],
  "cards_accepted": ["MasterCard", "Visa"],
  "open_days": ["Monday", "Wednesday"],
}

{
  "feature_id": "2",
  "services": ["Showers"],
  "cards_accepted": ["Visa"],
  "open_days": ["Tuesday", "Thursday"],
}

Spiders should not assume that feature 2 lacks Toilets, refuses to accept
MasterCard and is closed Monday, Wednesday and Friday-Sunday each week
unless there is other data available to confirm such facts.

Spiders should not assume that feature 1 is closed on Tuesday and
Thursday-Sunday each week unless there is other data available to
confirm such facts.

Spiders can confirm that feature 2 lacks Toilets if the front end website
displaying this data states for feature 2, "Toilets: not available".
That is, the ambiguity of missing data in the array is clarified by the
way in which the website has been programmed to interpret the data.