kaufland: fix invalid hours range interpretation of 00:00-00:00
More generally for other spiders, OpeningHours.add_range has been fixed
to ignore any open and close times which are identical for the same day.
This is particularly true when the open and close times are both 00:00
(or equivalents thereof). As the kaufland spider demonstrates, this
notation generally indicates the spider needs to use
OpeningHours.set_closed for such days.

Fixes #12989