storefinders/where2getit: add limit=10000 parameter
There has been (or is now) a limit parameter for the Where2GetIt
storefinder API method which was used, allowing a limit of up to 10000
results per query to be returned. Previously this wasn't observed to be
required for the ross_dress_for_less_us spider but more recently the
default limit for this brand appears to have been changed to 1000
results per query, versus the 1847 features this spider should be
expected to return. All 1847 features of this spider are again returned
if we manually set the limit to 10000 results per query.